requirements/base.txt
# you will need to install these requirements with `pip install -r requirements.txt`
PyYAML==3.13
google-auth==1.6.2
google-cloud-storage==1.19.1
python-dateutil==2.8.0
aiocron==1.3
funcy==1.12
gunicorn==19.9.0
Jinja2==2.10.1
requests==2.22.0
aiosmtplib==1.1.0
aiohttp==3.6.2
premailer==3.6.1
sentry-sdk==0.14.1
aiocontextvars==0.2.2
coloredlogs==10.0
git+https://github.com/Aidbox/aidbox-python-sdk.git@bb3011e40d91939f1dacfdf4f169ae542437a558
